requirements.txt
Flask
gunicorn
Werkzeug
Pillow
numpy
pandas
matplotlib
opencv-python
scikit-learn
seaborn
reportlab
